backend/requirements/local.txt
asgiref==3.8.1
attrs==25.3.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
psycopg[binary]==3.2.6 # https://github.com/psycopg/psycopg
#--test用に追加されたライブラリ--
pytest==8.3.5
pytest-xdist==3.6.1
pytest-sugar==1.0.0
pytest-mock==3.14.0
pytest-django==4.10.0
factory_boy==3.3.3